requirements.txt
pdfplumber>=0.9.0
openpyxl>=3.1.0
opencv-python-headless
pandas>=2.0.0
regex>=2023.0.0
pytesseract>=0.3.10
python-dotenv
pdf2image>=1.16.3
Pillow>=10.0.0
numpy
